+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Created two FATFS volumes on BoxLambda OS: the SD disk `sd0:`, and a RAM disk `ram:`.
 - Introduced FATFS FFI for Forth: `sw/components/forth/core/fs_ffi.cpp`.
 - Redirection of Picolibc's stdio to Forth's `emit` and `key`.
+- Re-added picorv_dma gw and sw components and projects in the `archived/` subdirectories.
 - Software components `sw/common/` containing common utility definitions suc
   as IN, OUT, INOUT parameters and die(...) macro.
 - Add added `forth/fs/` (forth filesystem) directory tree containing the BoxLambda Forth